Weapon and CharacterSkill

+++ Services/CharacterServices/CharacterService.cs
@@ -2,10 +2,12 @@
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Characters;
 using dotnet_rpg.Models;
+using Microsoft.AspNetCore.Http;
 using Microsoft.EntityFrameworkCore;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace dotnet_rpg.Services.CharacterServices
@@ -20,18 +22,22 @@ namespace dotnet_rpg.Services.CharacterServices
 
         private readonly DataContext _context;
         private readonly IMapper _mapper;
-        public CharacterService(IMapper mapper, DataContext context)
+        private readonly IHttpContextAccessor _httpContextAccessor;
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
         {
             _mapper = mapper;
             _context = context;
+            _httpContextAccessor = httpContextAccessor;
         }
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
         public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             ServiceResponse<List<GetCharacterDto>> ServiceResponse = new ServiceResponse<List<GetCharacterDto>>();
             Character character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(c => c.Id == GetUserId()); 
             await _context.Characters.AddAsync(character);
             await _context.SaveChangesAsync();
-            ServiceResponse.Data = (_context.Characters.Select(C => _mapper.Map<GetCharacterDto>(C))).ToList();
+            ServiceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(C => _mapper.Map<GetCharacterDto>(C))).ToList();
             return  ServiceResponse;
         }
 
@@ -40,11 +46,19 @@ namespace dotnet_rpg.Services.CharacterServices
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
             try
             {
-                Character character = await _context.Characters.FirstAsync(c => c.Id == id);
-                 _context.Characters.Remove(character);
-                await _context.SaveChangesAsync();
+                Character character = await _context.Characters.FirstAsync(c => c.Id == id && c.User.Id == GetUserId());
+                if(character != null)
+                {
+                    _context.Characters.Remove(character);
+                    await _context.SaveChangesAsync();
 
-                serviceResponse.Data = (_context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                    serviceResponse.Data = (_context.Characters.Where(c => c.User.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c))).ToList();
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Character not found.";
+                }
             }
             catch (Exception ex)
             {
@@ -54,10 +68,10 @@ namespace dotnet_rpg.Services.CharacterServices
             return serviceResponse;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> GetALlCharacter(int userId)
+        public async Task<ServiceResponse<List<GetCharacterDto>>> GetALlCharacter()
         {
             ServiceResponse<List<GetCharacterDto>> serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == userId).ToListAsync();
+            List<Character> dbCharacters = await _context.Characters.Where(c => c.User.Id == GetUserId()).ToListAsync();
             serviceResponse.Data = (dbCharacters.Select(C => _mapper.Map<GetCharacterDto>(C))).ToList();
             return serviceResponse;
         }
@@ -65,7 +79,7 @@ namespace dotnet_rpg.Services.CharacterServices
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             ServiceResponse<GetCharacterDto> serviceResponse = new ServiceResponse<GetCharacterDto>();
-            Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            Character dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
                                    
             return serviceResponse;
@@ -78,6 +92,9 @@ namespace dotnet_rpg.Services.CharacterServices
             try
             {
                 Character character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updateCharacterDto.Id);
+                if(character.User.Id == GetUserId())
+                {
+
                 character.Name = updateCharacterDto.Name;
                 character.Class = updateCharacterDto.Class;
                 character.Defence = updateCharacterDto.Defence;
@@ -89,6 +106,13 @@ namespace dotnet_rpg.Services.CharacterServices
                 await _context.SaveChangesAsync();
 
                 serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+
+                }
+                else
+                {
+                    serviceResponse.Success = false;
+                    serviceResponse.Message = "Not Found";
+                }
             }
             catch (Exception ex)
             {
